fix: clear incoming GRPC metadata from context
Etcd reads the incoming GRPC metadata (headers) in the context for authentication and possibly for other things. When an embedded etcd is used, these metadata can be coming from anywhere - since the etcd is embedded, it is received as-is by etcd as the incoming context. To prevent this, we make sure to clear the incoming metadata before hitting etcd.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>