fix: drain etcd Watch channels
Looking at the etcd `Watch` API, it seems that it always guarantees that
the returned channel is closed when the `Watch` is aborted/errors out.

At the same time, it doesn't look like `Watch` handler correctly handles
the case when context is canceled and nothing tries to receive on the
channel - it will still try to send there.

So adding two drain functions to make sure the etcd Watch API is
unblocked whenever we're done using it.

Changed the tests to use goroutine leak verifier to assure we are not
leaking anything.

I was able to confirm that the drain functions are in fact reading at
least a single message in our tests.

Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>